pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pcb2dlp"
version = "0.1.0"
description = "Convert PCB Gerber files to MSLA 3D printer formats for photoresist exposure"
requires-python = ">=3.10"
dependencies = [
    "gerbonara>=1.6",
    "cairosvg>=2.7",
    "Pillow>=10.0",
    "numpy>=1.24",
]

[project.scripts]
pcb2dlp = "pcb2dlp.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
